add peers to main opts
