+++ cmd/node/main.go
@@ -2,19 +2,19 @@ package main
 
 import (
 	"encoding/json"
-	pnet_node "github.com/amirylm/go-libp2p-pnet-node/lib"
-	"github.com/libp2p/go-libp2p"
-	connmgr "github.com/libp2p/go-libp2p-connmgr"
-	"github.com/libp2p/go-libp2p-core/crypto"
-	"github.com/libp2p/go-libp2p-core/host"
-	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/libp2p/go-libp2p-core/pnet"
 	"io/ioutil"
 	"log"
 	"os"
 	"os/signal"
 	"path/filepath"
 	"syscall"
+
+	pnet_node "github.com/amirylm/go-libp2p-pnet-node/lib"
+	"github.com/libp2p/go-libp2p"
+	connmgr "github.com/libp2p/go-libp2p-connmgr"
+	"github.com/libp2p/go-libp2p-core/crypto"
+	"github.com/libp2p/go-libp2p-core/peer"
+	"github.com/libp2p/go-libp2p-core/pnet"
 )
 
 func startNode(psk pnet.PSK, priv crypto.PrivKey, peers []peer.AddrInfo) (*pnet_node.PrivateNetNode, error) {
@@ -44,7 +44,7 @@ func main() {
 	priv, _, _ := crypto.GenerateKeyPair(crypto.Ed25519, 1)
 	psk := []byte("XVlBzgbaiCMRAjWwhTHctcuAxhxKQFDa")
 
-	var cfg Config
+	var cfg pnet_node.Options
 	p, err := filepath.Abs("./cmd/node/config.json")
 	check(err)
 	b, err := ioutil.ReadFile(p)
@@ -53,37 +53,25 @@ func main() {
 	}
 	log.Printf("num of peers: %d", len(cfg.Peers))
 
-	rel, err := startNode(psk, priv, cfg.Peers)
+	node, err := startNode(psk, priv, cfg.Peers)
 	check(err)
 
 	log.Println("node is ready:")
-	printHost(rel.Node)
+	log.Println(pnet_node.SerializePeer(node.Node))
 
 	// wait for a SIGINT or SIGTERM signal
 	ch := make(chan os.Signal, 1)
 	signal.Notify(ch, syscall.SIGINT, syscall.SIGTERM)
 	<-ch
 
-	errs := rel.Close()
+	errs := node.Close()
 	for _, err := range errs {
 		check(err)
 	}
 }
 
-type Config struct {
-	Peers []peer.AddrInfo `json:"peers"`
-}
-
 func check(err error) {
 	if err != nil {
 		log.Panic(err)
-	}
-}
-
-func printHost(h host.Host) {
-	id := h.ID().Pretty()
-	log.Printf("%s, listening on:", id)
-	for _, addr := range h.Addrs() {
-		log.Printf("\t- %s", addr.String())
 	}
 }
